+++ data/lib/domain/location_data_extension.dart
@@ -1,0 +1,8 @@
+import '../api/location/journey/journey.dart';
+import '../api/location/location.dart';
+
+extension LocationDataExtension on LocationData {
+  JourneyRoute toRoute() {
+    return JourneyRoute(latitude: latitude, longitude: longitude);
+  }
+}
